Update CepBrasil for .NET 7 and other improvements
Updated CepBrasil for .NET 7 compatibility and removed compatibility for .NET Core 3.0 and 3.1. Various other improvements were also made such as updating other dependencies, removing unused namespaces and unused properties in the .editorconfig, refactoring some code for better readability, modifying the async methods for better implementation, upgrading the package versions, and updating various test methods for better assertions using FluentAssertions. This update ensures the relevant and latest versions of frameworks and libraries are being used and improves the overall quality of the code and tests. The service error message was also updated to be more descriptive.